split out funcitons into a file each
