+++ packages/fake-data/src/populateDb/createInvoiceData.ts
@@ -1,0 +1,74 @@
+import type { Models } from '@tamanu/database';
+const { fake } = require('@tamanu/shared/test-helpers/fake');
+
+interface CreateInvoiceDataParams {
+  models: Models;
+  encounterId: string;
+  userId: string;
+  referenceDataId: string;
+  productId: string;
+}
+export const createInvoiceData = async ({
+  models: {
+    Invoice,
+    InvoiceDiscount,
+    InvoiceInsurer,
+    InvoicePayment,
+    InvoiceInsurerPayment,
+    InvoicePatientPayment,
+    InvoiceItemDiscount,
+    InvoiceItem,
+  },
+  encounterId,
+  userId,
+  referenceDataId,
+  productId,
+}: CreateInvoiceDataParams): Promise<void> => {
+  const invoice = await Invoice.create(
+    fake(Invoice, {
+      encounterId,
+    }),
+  );
+  await InvoiceDiscount.create(
+    fake(InvoiceDiscount, {
+      invoiceId: invoice.id,
+      appliedByUserId: userId,
+    }),
+  );
+  await InvoiceInsurer.create(
+    fake(InvoiceInsurer, {
+      invoiceId: invoice.id,
+      insurerId: referenceDataId,
+    }),
+  );
+  const invoicePayment = await InvoicePayment.create(
+    fake(InvoicePayment, {
+      invoiceId: invoice.id,
+    }),
+  );
+  await InvoiceInsurerPayment.create(
+    fake(InvoiceInsurerPayment, {
+      invoicePaymentId: invoicePayment.id,
+      insurerId: referenceDataId,
+    }),
+  );
+  await InvoicePatientPayment.create(
+    fake(InvoicePatientPayment, {
+      invoicePaymentId: invoicePayment.id,
+      methodId: referenceDataId,
+    }),
+  );
+  const invoiceItem = await InvoiceItem.create(
+    fake(InvoiceItem, {
+      invoiceId: invoice.id,
+      productId,
+      orderedByUserId: userId,
+    }),
+  );
+
+  await InvoiceItemDiscount.create(
+    fake(InvoiceItemDiscount, {
+      invoiceItemId: invoiceItem.id,
+    }),
+  );
+};
