+++ packages/fake-data/src/populateDb/index.ts
@@ -1,3 +1,15 @@
-export * from './helpers';
+// Specific groupings of data
 export * from './generateEachDataType';
 export * from './generateImportData';
+
+// Clinical data types
+export * from './createDbReportData';
+export * from './createTaskingData';
+export * from './createPatientData';
+export * from './createEncounterData';
+export * from './createInvoiceData';
+export * from './createImagingRequestData';
+export * from './createLabRequestData';
+export * from './createAppointmentData';
+export * from './createProgramRegistryData';
+export * from './createSurveyResponseData';
